LayoutTests/platform/wpe/imported/w3c/web-platform-tests/shadow-dom/focus/text-selection-with-delegatesFocus-expected.txt



FAIL shadow root has selectable text when focus is not delegated assert_greater_than: expected a number greater than 0 but got 0
FAIL shadow root has selectable text when focus is delegated assert_greater_than: expected a number greater than 0 but got 0
